+++ server/src/game/objects/building.ts
@@ -1,5 +1,5 @@
-import { BuildingCeilingDef, BuildingDef, BuildingObstacles } from "common/scripts/definitions/objects/buildings_base.ts";
-import { Angle, Hitbox2D, Stream, NullHitbox2D, Orientation, random, RotationMode, v2, Vec2, CheckpointContext } from "common/engine/core.ts";
+import { BuildingCeilingDef, BuildingDef, BuildingObstacles, BuildingPuzzleDef, PuzzleAction, PuzzleCondition } from "common/scripts/definitions/objects/buildings_base.ts";
+import { Angle, Hitbox2D, Stream, NullHitbox2D, Orientation, random, RotationMode, v2, Vec2, CheckpointContext, cloneDeep } from "common/engine/core.ts";
 import { StaticBody, StaticBodyPhysicalData } from "./static_body.ts";
 import { GameObjectType } from "common/scripts/others/constants.ts";
 import { type Obstacle } from "./obstacle.ts";
@@ -21,10 +21,208 @@ export class BuildingCeiling{
         return this.alive&&hb.colliding_with(this.hitbox)
     }
 }
+export class BuildingPuzzle{
+    parent:Building
+    id:string
+    global:boolean
+
+    code_value?:string
+    code_size?:number
+    current_code:string=""
+
+    complete_activated:boolean=false
+    complete_conditions?:PuzzleCondition[]
+    complete_actions?:PuzzleAction[]
+
+    fail_activated:boolean=false
+    fail_conditions?:PuzzleCondition[]
+    fail_actions?:PuzzleAction[]
+
+    check_actions?:PuzzleAction[]
+
+    current_actions:{actions:PuzzleAction[],tp:number}[]=[]
+
+    connections:Obstacle[]=[]
+
+    constructor(parent:Building,def:BuildingPuzzleDef){
+        this.parent=parent
+        this.id=def.idString??"main"
+        this.global=def.global??false
+        if(def.code){
+            this.code_size=def.code.size
+            this.code_value=def.code.value
+        }
+
+        this.complete_conditions=def.complete_conditions
+        this.complete_actions=def.complete_actions
+
+        this.fail_conditions=def.fail_conditions
+        this.fail_actions=def.fail_actions
+
+        this.check_actions=def.check_actions
+    }
+    get_object(id:number):Obstacle{
+        return this.parent.objects_ids[id]
+    }
+    input_piece(value:Obstacle) {
+        if(!(value.press_data?.activated||value.door_data))return
+        if(value.puzzle_value){
+            this.current_code+=value.puzzle_value??""
+            if(this.code_size!==undefined&&this.current_code.length>=this.code_size){
+                this.check()
+                this.current_code=""
+            }
+        }else{
+            this.check()
+        }
+    }
+    check() {
+        if(!this.complete_activated&&this.complete_conditions&&this.check_conditions(this.complete_conditions)){
+            this.complete_activated=true
+            if(this.complete_actions)this.start_actions(this.complete_actions,1)
+        }
+        if(!this.fail_activated&&this.fail_conditions&&this.check_conditions(this.fail_conditions)){
+            this.fail_activated=true
+            if(this.fail_actions)this.start_actions(this.fail_actions,2)
+        }
+        if(this.check_actions)this.start_actions(this.check_actions)
+    }
+    check_conditions(conditions: PuzzleCondition[]): boolean {
+        for (const c of conditions) {
+            let ok = false
+            switch (c.type){
+                case "code":{
+                    if(c.dont_need_orden){
+                        let code=""
+                        for(const o of this.connections){
+                            if(o.press_data?.activated&&o.puzzle_value){
+                                code+=o.puzzle_value
+                            }
+                        }
+                        ok=code===(c.value??this.code_value??"")
+                        if(c.negate)ok=!ok
+                    }else{
+                        ok=this.current_code===(c.value??this.code_value??"")
+                    }
+                    break
+                }
+                case "press":{
+                    const ids=typeof c.id==="number"?[c.id]:(c.id??[])
+                    ok=true
+                    for(const id of ids){
+                        const obj=this.get_object(id)
+                        if (!obj.press_data?.activated) {
+                            ok=false
+                            break
+                        }
+                    }
+                    break
+                }
+                case "break":{
+                    const ids=typeof c.id === "number"?[c.id]:(c.id??[])
+                    ok=true
+                    for(const id of ids) {
+                        const obj=this.get_object(id)
+                        if (!obj.health_data.dead) {
+                            ok = false
+                            break
+                        }
+                    }
+                    break
+                }
+            }
+            if(c.negate)ok=!ok
+            if(!ok)return false
+        }
+        return true
+    }
+    start_actions(actions:PuzzleAction[],tp:number=0){
+        this.current_actions.push({actions:cloneDeep(actions),tp})
+    }
+    tick(dt:number){
+        for(let i=0;i<this.current_actions.length;i++){
+            const a=this.current_actions[i].actions[0]
+            if(!a)continue
+            switch(a.type){
+                case "door":{
+                    this.current_actions[i].actions.shift()
+                    const objects:(Obstacle|undefined)[]=[]
+                    if(typeof a.id==="number"){
+                        objects.push(this.get_object(a.id))
+                    }else{
+                        for(const id of a.id){
+                            objects.push(this.get_object(id))
+                        }
+                    }
+                    for(const o of objects){
+                        if(!(o&&o.door_data))continue
+                        o.door_data.dirty=true
+                        o.set_dirty_part()
+                        if(a.open_state!==undefined)o.door_open(a.open_state)
+                        if(a.locked!==undefined)o.door_data.locked=a.locked
+                        if(a.only_side!==undefined)o.door_data.only_side=a.only_side
+                    }
+                    break
+                }
+                case "press":{
+                    this.current_actions[i].actions.shift()
+                    const objects:(Obstacle|undefined)[]=[]
+                    if(typeof a.id==="number"){
+                        objects.push(this.get_object(a.id))
+                    }else{
+                        for(const id of a.id){
+                            objects.push(this.get_object(id))
+                        }
+                    }
+                    for(const o of objects){
+                        if(!(o&&o.press_data))continue
+                        o.press_data.dirty=true
+                        o.set_dirty_part()
+                        if(a.activated!==undefined)o.press_data.activated=a.activated
+                        if(a.locked!==undefined)o.press_data.locked=a.locked
+                        if(a.can_switch!==undefined)o.press_data.allow_switch=a.can_switch
+                    }
+                    break
+                }
+                case "wait":{
+                    a.time-=dt
+                    if(a.time<=0){
+                        this.current_actions[i].actions.shift()
+                    }
+                    break
+                }
+                case "puzzle":{
+                    this.current_actions[i].actions.shift()
+                    const puzzle=a.id===undefined?this:this.parent.puzzles[a.id]
+                    if(a.lock!==undefined){
+                        if(a.lock){
+                            puzzle.complete_activated=true
+                            puzzle.fail_activated=true
+                        }else{
+                            puzzle.complete_activated=false
+                            puzzle.fail_activated=false
+                        }
+                    }
+                    break
+                }
+                case "check_fail":
+                    this.current_actions[i].actions.shift()
+                    break
+            }
+            if(this.current_actions[i].actions.length===0){
+                this.current_actions[i].actions.splice(i,1)
+                i--
+            }
+        }
+    }
+
+}
 export class Building extends StaticBody {
     override string_type = "building"
     override number_type = GameObjectType.Building
     def!: BuildingDef
+
+    puzzles:Record<string,BuildingPuzzle>={}
 
     physical_data:{
         dirty:boolean
@@ -47,6 +245,7 @@ export class Building extends StaticBody {
         passable_by_bullets:false,
         stairs:[]
     }
+
     children:BuildingObstacleChild[]=[]
     objects_ids:Record<number,Obstacle>={}
     ceilings:BuildingCeiling[]=[]
@@ -57,12 +256,17 @@ export class Building extends StaticBody {
         this.allow_net_update=true
     }
 
+    override tick(dt:number):void{
+        for(const p in this.puzzles){
+            this.puzzles[p].tick(dt)
+        }
+    }
     override on_net_update(): void {
         this.physical_data.dirty=false
     }
 
     set_definition(def: BuildingDef) {
-        if (this.def) return
+        if(this.def) return
         this.def = def
 
         if (def.hitbox){
@@ -75,6 +279,7 @@ export class Building extends StaticBody {
             this.physical_data.spawn_hitbox=this.physical_data.hitbox
         }
 
+        this.allow_tick=this.def.content.puzzles!==undefined
         this.physical_data.no_collision=this.def.no_collisions??false
         this.physical_data.no_bullets_collision=this.def.no_bullet_collision??false
         this.physical_data.reflect_bullets=this.def.reflect_bullets??false
@@ -119,6 +324,12 @@ export class Building extends StaticBody {
     }
     generate(position: Vec2){
         this.begin_generate(position)
+
+        for(const p of this.def.content.puzzles??[]){
+            const puzzle=new BuildingPuzzle(this,p)
+            this.puzzles[puzzle.id]=puzzle
+            if(p.global)this.game.puzzles[puzzle.id]=puzzle
+        }
 
         /*for(const f of this.def.floors??[]){
             const hb=f.hitbox.transform(this.position)
@@ -165,14 +376,23 @@ export class Building extends StaticBody {
                 }
             }
 
-            if(obj.door_data&&o.only_side){
-                obj.door_data.only_side=o.only_side
-            }
             if(obj.def.expanded_behavior?.type===4&&o.press_data&&obj.press_data){
                 if(o.press_data.activated!==undefined)obj.press_data.activated=o.press_data.activated
                 if(o.press_data.locked!==undefined)obj.press_data.activated=o.press_data.locked
                 if(o.press_data.allow_switch!==undefined)obj.press_data.allow_switch=o.press_data.allow_switch
             }
+            if(obj.door_data&&o.door_data){
+                if(o.door_data.open_state!==undefined)obj.door_open(o.door_data.open_state)
+                if(o.door_data.locked!==undefined)obj.door_data.locked=o.door_data.locked
+                if(o.door_data.only_side!==undefined)obj.door_data.only_side=o.door_data.only_side
+            }
+
+            if(o.puzzle_piece){
+                obj.puzzle=this.puzzles[o.puzzle_piece.id??"main"]
+                obj.puzzle_value=o.puzzle_piece.value
+                obj.puzzle.connections.push(obj)
+            }
+
             this.children.push({obj,def:o,type:0})
         }
         for (const b of this.def.content.sub_building ?? []) {